try podman binary

+++ packages/compass-e2e-tests/scripts/atlascli.sh
@@ -1,12 +1,13 @@
 #!/bin/bash
 
 declare ATLAS_URL
+declare PODMAN_URL
 declare PLATFORM
 declare ARCH
 
 ATLAS_VERSION=1.12.1
+PODMAN_VERSION=4.7.1
 TMP_PATH=../tmp
-ATLAS_CLI="${TMP_PATH}/atlas/bin/atlas"
 
 setup_environment() {
     if [[ $(uname) == Darwin ]]; then
@@ -17,6 +18,9 @@ setup_environment() {
             ARCH=arm64
         fi
         ATLAS_URL="https://fastdl.mongodb.org/mongocli/mongodb-atlas-cli_${ATLAS_VERSION}_${PLATFORM}_${ARCH}.zip"
+        
+        PODMAN_ARCH=[[ $ARCH = 'arm64' ]] && 'arm64' || 'amd64'
+        PODMAN_URL="https://github.com/containers/podman/releases/download/v${PODMAN_VERSION}/podman-remote-release-darwin_${PODMAN_ARCH}.zip"
     fi
 }
 
@@ -28,22 +32,35 @@ is_local_atlas_supported() {
     fi
 }
 
-setup_atlascli() {
-    echo "Downloading atlascli for $PLATFORM-$ARCH from $ATLAS_URL"
-
-    mkdir -p "$TMP_PATH"
-    curl -o "${TMP_PATH}/atlas.zip" "$ATLAS_URL"
-    unzip -q -o "${TMP_PATH}/atlas.zip" -d "${TMP_PATH}/atlas"
-
-    chmod +x "$ATLAS_CLI"
-    echo "Atlas CLI installed"
-    echo "$($ATLAS_CLI --version)"
-
-    rm "${TMP_PATH}/atlas.zip"
+setup_podman() {
+    if ! command -v podman &> /dev/null; then
+        echo "Downloading podman for $PLATFORM-$ARCH from $PODMAN_URL"
+        mkdir -p "$TMP_PATH"
+        curl -o "${TMP_PATH}/podman.zip" "$PODMAN_URL"
+        unzip -q -o "${TMP_PATH}/podman.zip" -d "${TMP_PATH}/podman"
+        rm "${TMP_PATH}/podman.zip"
+        export PATH=$PATH:"${TMP_PATH}/podman/usr/bin"
+    fi
+    echo "Podman installed"
+    echo "podman --version"
 }
 
-teardown_atlascli() {
+setup_atlascli() {
+    if ! command -v atlas &> /dev/null; then
+        echo "Downloading atlascli for $PLATFORM-$ARCH from $ATLAS_URL"
+        mkdir -p "$TMP_PATH"
+        curl -o "${TMP_PATH}/atlas.zip" "$ATLAS_URL"
+        unzip -q -o "${TMP_PATH}/atlas.zip" -d "${TMP_PATH}/atlas"
+        rm "${TMP_PATH}/atlas.zip"
+        export PATH=$PATH:"${TMP_PATH}/atlas/bin"
+    fi
+    echo "Atlas installed"
+    echo "atlas --version"
+}
+
+teardown() {
     rm -rf "${TMP_PATH}/atlas"
+    rm -rf "${TMP_PATH}/podman"
 }
 
 usage() {
@@ -68,13 +85,14 @@ if [ "$1" == "setup" ]; then
         exit 1
     fi
 
+    setup_podman
     setup_atlascli
 
     PORT=$2
     NAME=$3
 
     echo "Setting up a local deployment $NAME on port $PORT"
-    "$ATLAS_CLI" deployments setup "$NAME" --force --type LOCAL --port "$PORT"
+    atlas deployments setup "$NAME" --force --type LOCAL --port "$PORT"
     exit 0
 fi
 
@@ -86,8 +104,8 @@ if [ "$1" == "teardown" ]; then
     if [[ $(uname) == "Darwin" ]]; then
         NAME=$2
         echo "Tearing down a local deployment $NAME"
-        "$ATLAS_CLI" deployments delete "$NAME" --force
-        teardown_atlascli
+        atlas deployments delete "$NAME" --force
+        teardown
         exit 0
     else
         echo "Skipping atlascli teardown for: $(uname)"
